GroupClient
Currently, the GroupClientStart is meant to test things and see if they are working as intended or not. Currently, the disconnecting does not seem to implement correctly. But other than that, everything seems to be okay. Run the GroupServerStart first, then you will prompted to enter a username. Run the GroupClientStart on a seperate command window.