Reorganized classes
To run:
1. Compile all (java *.java)
2. Run GroupServerStart
3. Run Client

Eliminated GroupClientStart by using Client to interface with the user
(ask which client to run) and moving the GCS functions to GroupClient.
When the file server/client/thread are implemented, FileServerStart (an
analogous class to GroupServerStart) can be run to start the file
server, and then Client will run FileClient if the user so chooses.
Still need to make sure we can go back and forth between GroupClient
and FileClient easily.